- And one I missed
git-svn-id: https://dev.chiasoft.net:8443/svn/svn_vana/trunk@1917 23f3cdc0-e6d1-42be-a0c1-51f4125517bb

+++ src/ChannelServer/PlayerActiveBuffs.cpp
@@ -328,5 +328,5 @@ const bool PlayerActiveBuffs::hasHyperBody() {
 }
 
 const int32_t PlayerActiveBuffs::getHyperBody() {
-	return (getActiveSkillLevel(Jobs::Spearman::HyperBody) > 0 ? Jobs::Spearman::HyperBody : Jobs::SuperGM::HyperBody);
+	return (getActiveSkillLevel(Jobs::Spearman::HyperBody) > 0 ? (int32_t)Jobs::Spearman::HyperBody : (int32_t)Jobs::SuperGM::HyperBody);
 }
